Removed dynamoDB. Modularized display scores and read sprite sheet functions.
