+++ src/classes/clickableObject.js
